pyproject.toml
[project]
name = "mcp-cpu-usage"
version = "0.1.0"
description = "Get process information and CPU usage through MCP for an LLM."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.53.0",
    "mcp>=1.6.0",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
